requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.10.5
pydantic-settings==2.7.0
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.2.0